Support for DESTDIR and uninstall of application
I wanted to build the X11 version PDCurses and then distribute it to other VMs I have running. GNU documentation pointed me to DESTDIR in make, and a fair amount of testing got me to the attached Makefile.in with these changes. I ended up adding support for DESTDIR, and for uninstall (because I got tired of manually rm'ing all the files installed.) Is there anything else needed?